inject dependencies and class factories for formulas
