+++ application/controllers/admin/courses.php
@@ -1,5 +1,8 @@
 <?php
 
+use Application\Services\DependencyInjection\ContainerFactory;
+use Application\Services\Formula\Builder;
+
 /**
  * Courses controller for backend.
  *
@@ -484,6 +487,12 @@ class Courses extends LIST_Controller
     
     public function add_task_set_type(): void
     {
+        // formula builder
+        $container = ContainerFactory::getContainer();
+        /** @var Builder $formulaBuilder */
+        $formulaBuilder = $container->get(Builder::class);
+        $formula = $formulaBuilder->build("");
+        
         $this->load->library('form_validation');
         
         $url = $this->uri->ruri_to_assoc(3);
