+++ application/services/Formula/Node/Formula.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Application\Services\Formula\Node;
+
+/**
+ * @property Formula_node $left
+ * @property Formula_node $right
+ */
+
+interface Formula extends Formula_node
+{
+    public function get_left();
+    public function get_right();
+}
