+++ application/services/Formula/Node/Formula_node.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Application\Services\Formula\Node;
+
+interface Formula_node
+{
+    public function compute(): float;  // vrati null ak sa neda vyhodnotit
+    public function evaluate(): bool;
+    public function toString(): string;
+}
