--- application/services/formula/Formula_node.php
@@ -1,8 +1,0 @@
-<?php
-
-interface Formula_node
-{
-    public function compute();
-    public function evaluate();
-    public function toString();
-}
